requirements.txt
fitz
Pillow
pytesseract
PyMuPDF
Flask
